w250825

+++ Programm_32F407VE/inc/HEHAL.h
@@ -6,9 +6,10 @@
 #include "stm32f4xx.h"
 
 /*defaines*/
-#define ENABLE_BIT(reg, bit) ((reg) |= (bit)) 
-#define DISABLE_BIT(reg, bit) ((reg) &= ~(bit)) 
+#define ENABLE_BIT(reg, bit) ((reg) |= (uint32_t)(bit)) 
+#define DISABLE_BIT(reg, bit) ((reg) &= ~(uint32_t)(bit)) 
 #define Read_BIT(reg, bit)    ((reg) & (bit)) 
+
 #define Write_REG(REG, VAL)   ((REG) = (VAL))
 #define Read_REG(REG)         ((REG))
 #define Clear_REG(REG)        ((REG) = (0x0))
